Updated subpage links
This is a very small change that allows the website to have full site navigation (if that's what its called?). The user can now use the "Back" link to return to home instead of having to click the back arrow to return to the previous tab in their browser. I might replace this with a navbar, but I will replace the code for it when the time comes.